+++ main/src/ca/uwaterloo/flix/language/ast/Type.scala
@@ -1048,5 +1048,53 @@ object Type {
     case tpe => tpe
   }
 
+  /**
+    * Returns the Flix Type of a Java Class
+    */
+  def getFlixType(c: Class[_]): Type = {
+    if (c == java.lang.Boolean.TYPE) {
+      Type.Bool
+    }
+    else if (c == java.lang.Byte.TYPE) {
+      Type.Int8
+    }
+    else if (c == java.lang.Short.TYPE) {
+      Type.Int16
+    }
+    else if (c == java.lang.Integer.TYPE) {
+      Type.Int32
+    }
+    else if (c == java.lang.Long.TYPE) {
+      Type.Int64
+    }
+    else if (c == java.lang.Character.TYPE) {
+      Type.Char
+    }
+    else if (c == java.lang.Float.TYPE) {
+      Type.Float32
+    }
+    else if (c == java.lang.Double.TYPE) {
+      Type.Float64
+    }
+    else if (c == classOf[java.math.BigInteger]) {
+      Type.BigInt
+    }
+    else if (c == classOf[java.lang.String]) {
+      Type.Str
+    }
+    else if (c == java.lang.Void.TYPE) {
+      Type.Unit
+    }
+    // handle arrays of types
+    else if (c.isArray) {
+      val comp = c.getComponentType
+      val elmType = getFlixType(comp)
+      Type.mkArray(elmType, Type.False, SourceLocation.Unknown)
+    }
+    // otherwise native type
+    else {
+      Type.mkNative(c, SourceLocation.Unknown)
+    }
+  }
 
 }
